+++ CHANGELOG.md
@@ -6,6 +6,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [3.12.1908.0 - August 2019 release]
+
+### Added
+
+- Add extension to create team site with no group using SPSiteManager #2364 [NicolajHedeager]
+
+### Changed
+
+- Fix: Added Built in CT Html Page layout #2321 [SchauDK]
+- Fix: Handle GroupSiteManager/CreateGroupEx with SiteStatus = 1 (provisioning status)
+- Fix: Handle SPSiteManager/Create with SiteStatus = 1 (provisioning status)
+- Fix: Increase default timeout for ALM API calls
+
 ## [3.11.1907.0 - July 2019 release]
 
 ### Added